requirements.txt
# Core ML/RL Libraries
torch>=2.0.0
numpy>=1.21.0
matplotlib>=3.5.0
tensorboard>=2.10.0

# Game Environment
pygame>=2.1.0

# Configuration and Utilities
pyyaml>=6.0
tqdm>=4.64.0
wandb>=0.13.0

# Data Science and Analysis
pandas>=1.5.0
seaborn>=0.11.0
jupyter>=1.0.0
notebook>=6.4.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Development Tools
black>=22.0.0
flake8>=5.0.0
mypy>=0.991
